+++ CHANGELOG.md
@@ -5,10 +5,17 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Change field name `duration` to `last_time`.
+
 ## [0.1.0] - 2023-03-27
 
 ### Added
 
 - Move from giganto
 
+[Unreleased]: https://github.com/aicers/giganto-client/compare/0.1.0...main
 [0.1.0]: https://github.com/aicers/giganto-client/tree/0.1.0